before rewriting biToString

+++ 2538/berezhko/include/bigint.h
@@ -52,7 +52,7 @@ void biDivRem(BigInt *quotient, BigInt *remainder, BigInt numerator, BigInt deno
 /** Compare two BitInts.
  * \returns sign(a - b)
  */
-long long biCmp(BigInt a, BigInt b);
+int biCmp(BigInt a, BigInt b);
 
 #ifdef __cplusplus
 }
